flaskORM/requirements.txt
alembic==1.4.2
aniso8601==8.0.0
appdirs==1.4.4
asgiref==3.2.10
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
cryptography==2.9.2
DateTime==4.3
debtcollector==2.2.0
distlib==0.3.1
Django==3.1
djangorestframework==3.11.1
env==0.1.0
filelock==3.0.12
Flask==1.1.2
Flask-Migrate==2.5.3
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.4
idna==2.10
importlib-metadata==1.7.0
importlib-resources==3.0.0
iso8601==0.1.12
itsdangerous==1.1.0
Jinja2==2.11.2
keystoneauth1==4.2.1
Mako==1.1.3
MarkupSafe==1.1.1
msgpack==1.0.0
mysql==0.0.2
mysqlclient==2.0.1
netaddr==0.8.0
netifaces==0.10.9
os-service-types==1.7.0
oslo.config==8.3.1
oslo.i18n==5.0.0
oslo.serialization==4.0.0
oslo.utils==4.4.0
packaging==20.4
pbr==5.4.5
Pillow==7.2.0
positional==1.2.1
prettytable==0.7.2
pycparser==2.20
PyJWT==1.7.1
PyMySQL==0.10.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-editor==1.0.4
python-keystoneclient==2.3.2
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
rfc3986==1.4.0
six==1.15.0
SQLAlchemy==1.3.19
sqlparse==0.3.1
stevedore==3.2.0
urllib3==1.25.10
virtualenv==20.0.31
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
zope.interface==5.1.0
